Moved types folder to root of the project
