+++ app/conversations/[conversationId]/components/Body.tsx
@@ -3,7 +3,7 @@
 import axios from "axios";
 import { useEffect, useRef, useState } from "react";
 import useConversation from "@/hooks/useConversation";
-import { FullMessageType } from "@/app/types";
+import { FullMessageType } from "@/types";
 import MessageBox from "./MessageBox";
 import { pusherClient } from "@/app/libs/pusher";
 import { find } from "lodash";
